Add inline assets option
