+++ .gitignore
@@ -1,4 +1,5 @@
 .rvmrc
+.ruby-version
 .bundle
 
 ## MAC OS
