+++ Gemfile
@@ -16,8 +16,8 @@ end
 
 group :development do
   gem "rubocop"
+  gem "sass"
   # sprockets 4.0 requires ruby 2.5+
   gem "sprockets", "~> 3.7"
   gem "uglifier"
-  gem "yui-compressor"
 end
