+++ lib/simplecov/formatter/html_formatter/version.rb
@@ -1,0 +1,9 @@
+# frozen_string_literal: true
+
+module SimpleCov
+  module Formatter
+    class HTMLFormatter
+      VERSION = "0.12.3"
+    end
+  end
+end
