+++ views/layout.erb
@@ -6,7 +6,6 @@
     <script src='<%= assets_path('application.js') %>' type='text/javascript'></script>
     <link href='<%= assets_path('application.css') %>' media='screen, projection, print' rel='stylesheet' type='text/css' />
     <link rel="shortcut icon" type="image/png" href="<%= assets_path("favicon_#{coverage_css_class(result.source_files.covered_percent)}.png") %>" />
-    <link rel="icon" type="image/png" href="<%= assets_path('favicon.png') %>" />
   </head>
 
   <body<%= ' data-branch-coverage=true' if branchable_result? %>>
